Measure impact of thread count on metrics
Spreadsheet: https://docs.google.com/spreadsheets/d/1EFdgQopAMDR5zC36ZkXLlAebrI2skzpe155SgC1TCSc/edit?usp=sharing

The goal is to try to measure how much is gained or lost when increasing
or decreasing thread count, as to try to locate a decent tradeoff between
increased throughput (cost reduction), and reduced latency (user experience).

The application is configured as doing roughtly 50% IOs, with roughtly
100ms reponse time (in practice it's more like 115).

The latencies are measured in the Rack application side, as to ignore request
queuing. Queueing shouldn't happen in production aside from breif spikes of
traffic. If it does, it means the application is under-capacited, and capacity
should be increased, so not a scenario we're benchmarking for.

Notes.

This benchmark isn't trying to acheive maximum scientificity. It's
performed on a development machine, etc. However the number differences
are large enough that this relatively low benchmark precision is good
enough™ for the use case.